// app/benchmarking/components/enums.ts
export enum BenchmarkHomeViews {
  HOME,
  NEW_SESSION,
}

export enum BenchmarkNewSessionViews {
  PRIMARY_USE_CASE = 'PRIMARY_USE_CASE',
  TOPICS_SELECTION = 'TOPICS_SELECTION',
  RECOMMENDED_TESTS = 'RECOMMENDED_TESTS',
  ENDPOINTS_SELECTION = 'ENDPOINTS_SELECTION',
  NEW_ENDPOINT_FORM = 'NEW_ENDPOINT_FORM',
  EDIT_ENDPOINT_FORM = 'EDIT_ENDPOINT_FORM',
  COOKBOOKS_SELECTION = 'COOKBOOKS_SELECTION',
  BENCHMARK_RUN_FORM = 'BENCHMARK_RUN_FORM',
}

export enum GradingLevelsMlcEnum {
  LOW_RISK = 'Low Risk',
  MODERATE_LOW_RISK = 'Moderate-Low Risk',
  MODERATE_RISK = 'Moderate Risk',
  MODERATE_HIGH_RISK = 'Moderate-High Risk',
  HIGH_RISK = 'High Risk',
}